tourback/pyproject.toml
[project]
name = "tourback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "python-jose>=3.5.0",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
